+++ app/src/main/java/dev/digitaldreamweavers/qrguard/ui/camera/ViewFinderViewModel.java
@@ -43,11 +43,6 @@ public class ViewFinderViewModel extends ViewModel {
 
     private String TAG = "ViewFinderViewModel";
 
-    // Use provided MLKit BarcodeScanner
-    private BarcodeScanner qrScanner;
-
-    private ExecutorService executor = Executors.newSingleThreadExecutor();
-    private ImageCapture imageCapture;
 
     public LiveData<Integer> getFabStatus() {
         return fabStatus;
@@ -69,88 +64,6 @@ public class ViewFinderViewModel extends ViewModel {
     }
 
 
-    void bindPreview(@NonNull ProcessCameraProvider cameraProvider) {
-        cameraProvider.unbindAll();
-
-        Preview preview = new Preview.Builder()
-                .build();
-
-
-        CameraSelector cameraSelector = new CameraSelector.Builder()
-                //.requireLensFacing(CameraSelector.LENS_FACING_BACK)
-                .build();
-
-        Log.w(TAG, "Setting Surface Provider for camera...");
-        preview.setSurfaceProvider(qrViewFinder.getSurfaceProvider());
-
-        ImageAnalysis imageAnalyser = buildBarcodeScanner();
-        Log.i(TAG, "Image Analyser attached.");
-
-
-        Camera camera = cameraProvider.bindToLifecycle(this, cameraSelector, preview, imageAnalyser);
-    }
-
-
-    // buildBarcodeScanner: Returns an image analyser to be bind to a camera.
-    @OptIn(markerClass = ExperimentalGetImage.class)
-    private ImageAnalysis buildBarcodeScanner() {
-        Log.i(TAG, "Building analyser...");
-        ImageAnalysis imageAnalysis = new ImageAnalysis.Builder()
-                .setBackpressureStrategy(ImageAnalysis.STRATEGY_KEEP_ONLY_LATEST)
-                .build();
-
-        Log.i(TAG, "Setting analyser...");
-        imageAnalysis.setAnalyzer(executor, image -> {
-            // Log.i(TAG, "Analysing...");
-            InputImage inputImage = InputImage.fromMediaImage(Objects.requireNonNull(image.getImage()), image.getImageInfo().getRotationDegrees());
-            // Use BarcodeScanner to scan the image
-            qrScanner.process(inputImage)
-                    .addOnSuccessListener(barcodes -> {
-                        if (barcodes.isEmpty()) {
-                            //Log.i(TAG, "No barcode detected.");
-                            updateFAB(0);
-                        } else {
-                            for (Barcode barcode : barcodes) {
-
-                                try {
-                                    URL barcodeURL = new URL(barcode.getRawValue());
-                                    Log.i(TAG, "Valid URL detected, now scanning: " + barcodeURL.toString());
-                                    updateFAB(2);
-                                } catch (Exception e) {
-                                    Log.i(TAG, "Not a valid URL: " + barcode.getRawValue());
-                                    updateFAB(1);
-                                }
-                            }
-                        }
-                    })
-                    .addOnFailureListener(e -> {
-                        Log.w(TAG, "Barcode scanning failed: ", e);
-                    })
-                    .addOnCompleteListener(result -> {
-                        image.close();
-                        //Log.i(TAG, "Image closed.");
-                    });
-        });
-
-        return imageAnalysis;
-    }
-
-    private void startCamera() {
-
-        this.getCameraProviderFuture().observe(this, cameraProviderFuture -> {
-            if (cameraProviderFuture != null) {
-                try {
-                    ProcessCameraProvider cameraProvider = cameraProviderFuture.get();
-                    bindPreview(cameraProvider);
-                } catch (ExecutionException | InterruptedException e) {
-                    Log.w(TAG, "Could not add Camera Listener: ", e);
-                }
-            } else {
-                Log.w(TAG, "Camera Provider Future is null");
-            }
-        });
-
-    }
 
 
 }
